refactor(datastore): schemaディレクトリの冗長なネストを解消
src/drizzle-orm/schema/ を src/schema/ にフラット化し、公開パスを
@trend-diary/datastore/schema に短縮する。drizzle-orm/ 層は schema/ のみを
含む冗長な階層で、かつライブラリ名 drizzle-orm と同名のため import 元が
ライブラリか自パッケージか判別しにくかった。
