+++ application/apps/cron/src/feed/article-store.test.ts
@@ -1,5 +1,5 @@
 import type { D1Database } from '@cloudflare/workers-types'
-import { articles } from '@trend-diary/datastore/drizzle-orm/schema'
+import { articles } from '@trend-diary/datastore/schema'
 import { afterAll, beforeEach, describe, expect, it } from 'vitest'
 import TEST_ENV from '../test-helper/env'
 import { countArticles, testRdb as db } from '../test-helper/rdb'
