+++ application/apps/cron/src/feed/fetch-articles.test.ts
@@ -1,5 +1,5 @@
 import Logger from '@trend-diary/common/logger'
-import { articles } from '@trend-diary/datastore/drizzle-orm/schema'
+import { articles } from '@trend-diary/datastore/schema'
 import { env } from 'cloudflare:test'
 import { eq } from 'drizzle-orm'
 import { afterAll, beforeEach, describe, expect, it, vi } from 'vitest'
