+++ application/apps/web/src/server/article/handler/get-articles.test.ts
@@ -1,6 +1,6 @@
 import { faker } from '@faker-js/faker'
 import { ServerError } from '@trend-diary/common/errors'
-import { articles, readHistories } from '@trend-diary/datastore/drizzle-orm/schema'
+import { articles, readHistories } from '@trend-diary/datastore/schema'
 import type * as ArticleModule from '@trend-diary/domain/article'
 import { createArticleUseCase } from '@trend-diary/domain/article'
 import { inArray } from 'drizzle-orm'
