+++ application/apps/web/src/test/helper/rdb.ts
@@ -1,5 +1,5 @@
-import * as schema from '@trend-diary/datastore/drizzle-orm/schema'
 import type { RdbClient } from '@trend-diary/datastore/rdb'
+import * as schema from '@trend-diary/datastore/schema'
 import { drizzle } from 'drizzle-orm/d1'
 import { platformEnv } from '@/test/setup/platform-proxy'
