+++ application/e2e/src/fixtures.ts
@@ -1,6 +1,6 @@
 import { test as base } from '@playwright/test'
-import * as schema from '@trend-diary/datastore/drizzle-orm/schema'
 import type { RdbClient } from '@trend-diary/datastore/rdb'
+import * as schema from '@trend-diary/datastore/schema'
 import { drizzle } from 'drizzle-orm/d1'
 import { openTestD1 } from './d1'
