+++ application/packages/datastore/package.json
@@ -5,7 +5,7 @@
   "type": "module",
   "exports": {
     "./rdb": "./src/rdb/index.ts",
-    "./drizzle-orm/schema": "./src/drizzle-orm/schema/index.ts",
+    "./schema": "./src/schema/index.ts",
     "./*": "./src/*.ts"
   },
   "scripts": {
