+++ application/packages/datastore/src/rdb/client.ts
@@ -1,6 +1,6 @@
 import type { D1Database } from '@cloudflare/workers-types'
 import { type DrizzleD1Database, drizzle as drizzleD1 } from 'drizzle-orm/d1'
-import * as schema from '../drizzle-orm/schema'
+import * as schema from '../schema'
 import { queryLogger } from './logger'
 
 export type RdbClient = DrizzleD1Database<typeof schema>
