+++ application/packages/datastore/src/schema/article.ts
