+++ application/packages/domain/src/account/infrastructure/command-impl.ts
@@ -1,8 +1,8 @@
 import { ServerError } from '@trend-diary/common/errors'
 import Logger from '@trend-diary/common/logger'
-import { activeUsers, users } from '@trend-diary/datastore/drizzle-orm/schema'
 import type { RdbClient } from '@trend-diary/datastore/rdb'
 import { wrapDbCall } from '@trend-diary/datastore/rdb'
+import { activeUsers, users } from '@trend-diary/datastore/schema'
 import { eq } from 'drizzle-orm'
 import { err, ok, type Result } from 'neverthrow'
 import type { Command, Notifier } from '../port'
