+++ application/packages/domain/src/account/infrastructure/mapper.ts
@@ -1,5 +1,5 @@
-import type { ActiveUser as RdbActiveUser } from '@trend-diary/datastore/drizzle-orm/schema'
 import { fromDbId } from '@trend-diary/datastore/rdb/id'
+import type { ActiveUser as RdbActiveUser } from '@trend-diary/datastore/schema'
 import type { CurrentUser } from '../schema/active-user-schema'
 
 export function mapToActiveUser(activeUser: RdbActiveUser): CurrentUser {
