+++ application/packages/domain/src/account/infrastructure/query-impl.ts
@@ -1,9 +1,9 @@
 import { ServerError } from '@trend-diary/common/errors'
 import type { Nullable } from '@trend-diary/common/types/utility'
-import { activeUsers } from '@trend-diary/datastore/drizzle-orm/schema'
 import type { RdbClient } from '@trend-diary/datastore/rdb'
 import { wrapDbCall } from '@trend-diary/datastore/rdb'
 import { toDbId } from '@trend-diary/datastore/rdb/id'
+import { activeUsers } from '@trend-diary/datastore/schema'
 import { eq } from 'drizzle-orm'
 import { err, ok, type Result } from 'neverthrow'
 import type { Query } from '../port'
