+++ application/packages/domain/src/article/infrastructure/command-impl.ts
@@ -1,12 +1,8 @@
 import { NotFoundError, ServerError } from '@trend-diary/common/errors'
-import {
-  articles,
-  normalizeDateTime,
-  readHistories,
-} from '@trend-diary/datastore/drizzle-orm/schema'
 import type { RdbClient } from '@trend-diary/datastore/rdb'
 import { wrapDbCall } from '@trend-diary/datastore/rdb'
 import { fromDbId, toDbId } from '@trend-diary/datastore/rdb/id'
+import { articles, normalizeDateTime, readHistories } from '@trend-diary/datastore/schema'
 import { and, eq, exists, sql } from 'drizzle-orm'
 import { err, ok, type Result } from 'neverthrow'
 import type { Command } from '../port'
