+++ application/packages/domain/src/article/infrastructure/query-impl.ts
@@ -3,10 +3,10 @@ import { addJstDays, toJstDate } from '@trend-diary/common/locale/date'
 import type { OffsetPaginationResult } from '@trend-diary/common/pagination'
 import { DEFAULT_LIMIT, DEFAULT_PAGE } from '@trend-diary/common/pagination'
 import type { Nullable } from '@trend-diary/common/types/utility'
-import { articles, normalizeDateTime } from '@trend-diary/datastore/drizzle-orm/schema'
 import type { RdbClient } from '@trend-diary/datastore/rdb'
 import { wrapDbCall } from '@trend-diary/datastore/rdb'
 import { fromDbId, toDbId } from '@trend-diary/datastore/rdb/id'
+import { articles, normalizeDateTime } from '@trend-diary/datastore/schema'
 import { eq, type SQL, sql } from 'drizzle-orm'
 import { err, ok, type Result } from 'neverthrow'
 import { ARTICLE_MEDIA, type ArticleMedia, assertArticleMedia } from '../media'
